requirements.txt
fastapi
uvicorn
pymongo
motor
tensorflow
numpy
scikit-learn
pillow
python-dotenv
